+++ core/indexer/tests/database.rs
@@ -25,7 +25,7 @@ async fn test_database() -> Result<()> {
     let config = Config::try_parse()?;
     let client = Client::new_from_config(&config)?;
     let height = 800000;
-    let hash = client.get_block_hash(height).await?;
+    let hash = client.get_block_hash(height as u64).await?;
     let block = BlockRow { height, hash };
 
     let (reader, writer, _temp_dir) = new_test_db(&config).await?;
@@ -52,7 +52,7 @@ async fn test_transaction() -> Result<()> {
     let tx = writer.connection().transaction().await?;
     let height = 800000;
     let client = Client::new_from_config(&config)?;
-    let hash = client.get_block_hash(height).await?;
+    let hash = client.get_block_hash(height as u64).await?;
     let block = BlockRow { height, hash };
     insert_block(&tx, block).await?;
     assert!(select_block_latest(&tx).await?.is_some());
@@ -92,7 +92,7 @@ async fn test_contract_state_operations() -> Result<()> {
 
     // Insert a transaction for the contract state
     let tx = TransactionRow::builder()
-        .height(height as i64)
+        .height(height)
         .txid("abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890".to_string())
         .tx_index(0)
         .build();
@@ -106,7 +106,7 @@ async fn test_contract_state_operations() -> Result<()> {
     let contract_state = ContractStateRow::builder()
         .contract_id(contract_id)
         .tx_id(tx_id)
-        .height(height as i64)
+        .height(height)
         .path(path.to_string())
         .value(value.clone())
         .build();
@@ -145,7 +145,7 @@ async fn test_contract_state_operations() -> Result<()> {
     assert_eq!(retrieved_state.value, value);
     assert_eq!(retrieved_value.unwrap(), value);
     assert!(!retrieved_state.deleted);
-    assert_eq!(retrieved_state.height, height as i64);
+    assert_eq!(retrieved_state.height, height);
     assert_eq!(retrieved_state.tx_id, tx_id);
 
     // Test with a newer version of the same contract state
@@ -159,7 +159,7 @@ async fn test_contract_state_operations() -> Result<()> {
 
     let txid2 = "fedcba0987654321fedcba0987654321fedcba0987654321fedcba0987654321";
     let tx2 = TransactionRow::builder()
-        .height(height2 as i64)
+        .height(height2)
         .txid(txid2.to_string())
         .tx_index(2)
         .build();
@@ -169,7 +169,7 @@ async fn test_contract_state_operations() -> Result<()> {
     let updated_contract_state = ContractStateRow::builder()
         .contract_id(contract_id)
         .tx_id(tx_id2)
-        .height(height2 as i64)
+        .height(height2)
         .path(path.to_string())
         .value(updated_value.clone())
         .build();
@@ -182,12 +182,12 @@ async fn test_contract_state_operations() -> Result<()> {
     let latest_value = get_latest_contract_state_value(&conn, contract_id, path)
         .await?
         .unwrap();
-    assert_eq!(latest_state.height, height2 as i64);
+    assert_eq!(latest_state.height, height2);
     assert_eq!(latest_state.value, updated_value);
     assert_eq!(latest_value, updated_value);
 
     // Delete the contract state
-    let deleted = delete_contract_state(&conn, height2 as i64, tx_id, contract_id, path).await?;
+    let deleted = delete_contract_state(&conn, height2, tx_id, contract_id, path).await?;
     assert!(deleted);
 
     let count = conn
@@ -223,17 +223,17 @@ async fn test_transaction_operations() -> Result<()> {
     insert_block(&conn, block).await?;
 
     let tx1 = TransactionRow::builder()
-        .height(height as i64)
+        .height(height)
         .txid("abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890".to_string())
         .tx_index(0)
         .build();
     let tx2 = TransactionRow::builder()
-        .height(height as i64)
+        .height(height)
         .txid("123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef0".to_string())
         .tx_index(1)
         .build();
     let tx3 = TransactionRow::builder()
-        .height(height as i64)
+        .height(height)
         .txid("fedcba0987654321fedcba0987654321fedcba0987654321fedcba0987654321".to_string())
         .tx_index(2)
         .build();
@@ -248,7 +248,7 @@ async fn test_transaction_operations() -> Result<()> {
     let tx1 = get_transaction_by_id(&conn, tx_id1).await?.unwrap();
     assert_eq!(tx1.id, Some(tx_id1));
     assert_eq!(tx1.txid, tx1.txid);
-    assert_eq!(tx1.height, height as i64);
+    assert_eq!(tx1.height, height);
 
     // Test get_transaction_by_txid
     let tx2 = get_transaction_by_txid(&conn, tx2.txid.as_str())
@@ -256,10 +256,10 @@ async fn test_transaction_operations() -> Result<()> {
         .unwrap();
     assert_eq!(tx2.id, Some(tx_id2));
     assert_eq!(tx2.txid, tx2.txid);
-    assert_eq!(tx2.height, height as i64);
+    assert_eq!(tx2.height, height);
 
     // Test get_transactions_at_height
-    let txs_at_height = get_transactions_at_height(&conn, height as i64).await?;
+    let txs_at_height = get_transactions_at_height(&conn, height).await?;
     assert_eq!(txs_at_height.len(), 3);
 
     // Verify all transactions are included - now using TransactionRow objects
@@ -288,7 +288,7 @@ async fn test_transaction_operations() -> Result<()> {
     insert_block(&conn, block2).await?;
 
     let tx4 = TransactionRow::builder()
-        .height(height2 as i64)
+        .height(height2)
         .txid("aabbccddeeff00112233445566778899aabbccddeeff00112233445566778899".to_string())
         .tx_index(0)
         .build();
@@ -296,17 +296,17 @@ async fn test_transaction_operations() -> Result<()> {
     let tx_id4 = insert_transaction(&conn, tx4).await?;
 
     // Verify get_transactions_at_height returns only transactions at the specified height
-    let txs_at_height1 = get_transactions_at_height(&conn, height as i64).await?;
+    let txs_at_height1 = get_transactions_at_height(&conn, height).await?;
     assert_eq!(txs_at_height1.len(), 3);
 
-    let txs_at_height2 = get_transactions_at_height(&conn, height2 as i64).await?;
+    let txs_at_height2 = get_transactions_at_height(&conn, height2).await?;
     assert_eq!(txs_at_height2.len(), 1);
 
     // Check the transaction details
     let tx4 = &txs_at_height2[0];
     assert_eq!(tx4.id, Some(tx_id4));
     assert_eq!(tx4.txid, tx4.txid);
-    assert_eq!(tx4.height, height2 as i64);
+    assert_eq!(tx4.height, height2);
 
     Ok(())
 }
